Weapons Trace From Camera instead of gunsocket.
Hitscan weapons now fire the trace (damage trace) from the camera instead of the muzzle socket of the gun. Ensure that bullet tracers (both 1p and 3p) spawn with a rotation towards the new end trace point based off of the new trace from location or (new) if an object is hit by the damage trace, the tracers are spawned wit their rotation set towards the hit location (previously was not neccessary since both the damage trace and *at least the 1p tracer* were done from the same location and followed the same path).